# docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Local-first OpenAI-compatible LLM gateway with model discovery, caching, and usage tracking.",
        "title": "thinroute API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/control/v1/cache/overview": {
            "get": {
                "tags": [
                    "control"
                ],
                "summary": "Get cached-only usage overview",
                "parameters": [
                    {
                        "description": "Number of days (default 30)",
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Grouping interval: daily, weekly, monthly, yearly (default daily)",
                        "name": "interval",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by exact model name",
                        "name": "model",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by provider name or provider type",
                        "name": "provider",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by request label (exact match)",
                        "name": "label",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by tracked user path subtree",
                        "name": "user_path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Cache mode filter: uncached, cached, all (cache overview always uses cached mode)",
                        "name": "cache_mode",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/usage.CacheOverview"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/control/v1/usage/daily": {
            "get": {
                "tags": [
                    "control"
                ],
                "summary": "Get usage breakdown by period",
                "parameters": [
                    {
                        "description": "Number of days (default 30)",
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Grouping interval: daily, weekly, monthly, yearly (default daily)",
                        "name": "interval",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by exact model name",
                        "name": "model",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by provider name or provider type",
                        "name": "provider",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by request label (exact match)",
                        "name": "label",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by tracked user path subtree",
                        "name": "user_path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Cache mode filter: uncached, cached, all (default uncached)",
                        "name": "cache_mode",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/usage.DailyUsage"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/control/v1/usage/labels": {
            "get": {
                "description": "Returns per-label token and cost aggregates. Requests carrying\nseveral labels count once per label, so rows overlap and do\nnot sum to the period totals. Unlabelled requests are omitted.",
                "tags": [
                    "control"
                ],
                "summary": "Get usage breakdown by request label",
                "parameters": [
                    {
                        "description": "Number of days (default 30)",
                        "name": "days",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "description": "Start date (YYYY-MM-DD)",
                        "name": "start_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "End date (YYYY-MM-DD)",
                        "name": "end_date",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by exact model name",
                        "name": "model",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by provider name or provider type",
                        "name": "provider",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by request label (exact match)",
                        "name": "label",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Filter by tracked user path subtree",
                        "name": "user_path",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Cache mode filter: uncached, cached, all (default uncached)",
                        "name": "cache_mode",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/usage.LabelUsage"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
        "/control/v1/usage/throughput": {
            "get": {
                "description": "Returns a fixed, trailing window of token-volume buckets\n(input / output / prompt-cached / locally-cached) at the\nrequested granularity, for the overview live chart.",
                "tags": [
                    "control"
                ],
                "summary": "Get the live token-throughput window",
                "parameters": [
                    {
                        "description": "Bucket granularity: second, minute, hour, day",
                        "name": "granularity",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/usage.TokenThroughput"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/core.GatewayError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "servers": [
        {
            "url": "/"
        }
    ],
    "components": {
...
        "schemas": {
...
            "core.GatewayError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "nullable": true
                    },
                    "message": {
                        "type": "string"
                    },
                    "param": {
                        "type": "string",
                        "nullable": true
                    },
                    "provider": {
                        "type": "string"
                    },
                    "status_code": {
                        "type": "integer"
                    },
                    "type": {
                        "$ref": "#/components/schemas/core.ErrorType"
                    }
                }
            },
...
            "usage.CacheOverview": {
                "type": "object",
                "properties": {
                    "daily": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/usage.CacheOverviewDaily"
                        }
                    },
                    "summary": {
                        "$ref": "#/components/schemas/usage.CacheOverviewSummary"
                    }
                }
            },
...
            "usage.DailyUsage": {
                "type": "object",
                "properties": {
                    "cache_write_input_tokens": {
                        "type": "integer"
                    },
                    "cached_input_tokens": {
                        "type": "integer"
                    },
                    "date": {
                        "type": "string"
                    },
                    "input_cost": {
                        "type": "number"
                    },
                    "input_tokens": {
                        "type": "integer"
                    },
                    "output_cost": {
                        "type": "number"
                    },
                    "output_tokens": {
                        "type": "integer"
                    },
                    "requests": {
                        "type": "integer"
                    },
                    "total_cost": {
                        "type": "number"
                    },
                    "total_tokens": {
                        "type": "integer"
                    },
                    "uncached_input_tokens": {
                        "description": "Provider prompt-cache split of the period's input, folded per row from\nraw_data (same source as the summary). Zero when the storage layer does\nnot populate them.",
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}